requirements.txt
# Python dependencies for DMS POC scripts
psycopg2-binary>=2.9.9
boto3>=1.34.0
